Rules to movement and a death mechanic
Howdy bois, so i've returned from my camping trip (yahoo) boutta fix this broken shit program and make it like, work or smth. I've added the rule to movement so that you're not able to 180 anymore (like in every other snake game), and i've added a simple forloop to test if the head of the snake is inside the body, so that's a death mechanic.
I'm considering making walls a death aswell, but i dont know yet, as i gotta test shit with the ai if it's able to do shit with this now.